Code refactoring of commons-database
