+++ libs/commons-database/src/main/kotlin/gov/cdc/ocio/database/models/Aggregation.kt
@@ -1,0 +1,13 @@
+package gov.cdc.ocio.database.models
+
+import com.google.gson.annotations.SerializedName
+
+/**
+ * Issue leve; of Report-ERROR OR WARNING
+ */
+enum class Aggregation {
+    @SerializedName("SINGLE")
+    SINGLE,
+    @SerializedName("BATCH")
+    BATCH
+}
