+++ libs/commons-database/src/main/kotlin/gov/cdc/ocio/database/models/Level.kt
@@ -1,0 +1,13 @@
+package gov.cdc.ocio.database.models
+
+import com.google.gson.annotations.SerializedName
+
+/**
+ * Issue level of Report-ERROR OR WARNING
+ */
+enum class Level {
+    @SerializedName("ERROR")
+    ERROR,
+    @SerializedName("WARNING")
+    WARNING
+}
